requirements.txt
numpy==1.21.0
pandas==1.3.0
matplotlib==3.4.3
scipy==1.7.0
tqdm==4.62.1
scikit-learn==0.24.0
statsmodels==0.12.2
seaborn==0.11.2
scikit-learn-extra
python-dotenv==0.19.2
pysolar==0.10
requests==2.26.0
cufflinks-py==0.17.3
plotly==5.4.0
scikit-image==0.18.3  # For the ternary plot in plotly
pvlib-python==0.9.0
# awscli==1.22.21  # For the HPC in Amazon EC2
# aws-parallelcluster==2.11.3  # For the HPC in Amazon EC2
